+++ SadafBI/SeperateTransactionInformation/Models/SqlSeperateTransactionModel.cs
@@ -31,7 +31,7 @@ namespace SadafBI.SeperateTransactionInformation.Models
             public int? depositCo { get; set; }
             public int? itManagement { get; set; }
             public int? facility { get; set; }
-            public int? tax { get; set; }
+            public long? tax { get; set; }
             public int? branchId { get; set; }
             public string branchName { get; set; } = "";
             public string bourseReference { get; set; } = "";
@@ -41,6 +41,15 @@ namespace SadafBI.SeperateTransactionInformation.Models
             public string postTradeTime { get; set; } = "";
             public int? cbranchId { get; set; }
             public string cbranchName { get; set; } = "";
-        }
+
+
+
+  
+     
+
+    
+   
+
+    }
     }
